Redesign quantum kernel model folder (#168)
* change: qunatum kernel config move to qkernel folder

* add: model type

* fix: test config

* fix: import source by version

* fix: global setting

* change: model folder stracture
