+++ configs/qkernel_models/template-generate-regression.yaml
@@ -2,6 +2,7 @@ description: "This is a template configuration file for the QXMT framework."
 
 global_settings:
   random_seed: &global_seed 42
+  model_type: "qkernel"
   task_type: "regression"
 
 dataset:
