+++ qxmt/models/qkernels/qrigge.py
@@ -9,8 +9,8 @@ from sklearn.model_selection import cross_val_score
 
 from qxmt.constants import DEFAULT_N_JOBS
 from qxmt.kernels.base import BaseKernel
-from qxmt.models.base import BaseKernelModel
 from qxmt.models.hyperparameter_search.search import HyperParameterSearch
+from qxmt.models.qkernels.base import BaseKernelModel
 
 
 class QRiggeRegressor(BaseKernelModel):
